requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
langchain-groq
Groq
faiss-cpu
streamlit
pypdf
python-dotenv
pandas
numpy
requests
tqdm
langchain-community
ipywidgets
python-dotenv
beautifulsoup4
langchain-huggingface
pdfplumber
matplotlib
seaborn
tqdm




